[refactor] DeliveryRouteRecord 분리
