+++ delivery-service/src/main/java/com/da2jobu/deliveryservice/application/DeliveryServiceImpl.java
@@ -1,0 +1,4 @@
+package com.da2jobu.deliveryservice.application;
+
+public class DeliveryServiceImpl {
+}
