+++ delivery-service/src/main/java/com/da2jobu/deliveryservice/application/dto/delivery/request/CreateDeliveryRequestDto.java
@@ -1,0 +1,4 @@
+package com.da2jobu.deliveryservice.application.dto.delivery.request;
+
+public class CreateDeliveryRequestDto {
+}
