--- delivery-service/src/main/java/com/da2jobu/deliveryservice/application/dto/response/GetDeliveriesResponseDto.java
@@ -1,4 +1,0 @@
-package com.da2jobu.deliveryservice.application.dto.response;
-
-public class GetDeliveriesResponseDto {
-}
