--- delivery-service/src/main/java/com/da2jobu/deliveryservice/domain/deliveryRouteRecord/DeliveryRouteRecordRepository.java
@@ -1,9 +1,0 @@
-package com.da2jobu.deliveryservice.domain.deliveryRouteRecord;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface DeliveryRouteRecordRepository extends JpaRepository<DeliveryRouteRecord, UUID> {
-
-}
