--- delivery-service/src/main/java/com/da2jobu/deliveryservice/domain/deliveryRouteRecord/DeliveryRouteStatus.java
@@ -1,7 +1,0 @@
-package com.da2jobu.deliveryservice.domain.deliveryRouteRecord;
-
-public enum DeliveryRouteStatus {
-    WAITING,   // 허브 이동 대기중
-    MOVING,    // 허브 이동중
-    ARRIVED    // 목적지 허브 도착
-}
