--- delivery-service/src/main/java/com/da2jobu/deliveryservice/domain/deliveryRouteRecord/RouteLocationType.java
@@ -1,6 +1,0 @@
-package com.da2jobu.deliveryservice.domain.deliveryRouteRecord;
-
-public enum RouteLocationType {
-    COMPANY,
-    HUB
-}
