--- delivery-service/src/main/java/com/da2jobu/deliveryservice/presentation/DeliveryRouteRecordController.java
@@ -1,4 +1,0 @@
-package com.da2jobu.deliveryservice.presentation;
-
-public class DeliveryRouteRecordController {
-}
